requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
humanize==4.5.0
idna==3.4
keyboard==0.13.5
markdown-it-py==2.1.0
mdurl==0.1.2
newsdataapi==0.1.6
psycopg2==2.9.5
pycparser==2.21
PyGithub==1.57
Pygments==2.14.0
PyNaCl==1.5.0
python-decouple==3.7
python-gitlab==3.13.0
pytz==2022.7
requests==2.28.2
requests-toolbelt==0.10.1
reviews==0.6.0
rich==13.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
